+++ cloudinary-http44/src/main/java/com/cloudinary/http44/UploaderStrategy.java
@@ -99,7 +99,7 @@ public class UploaderStrategy extends AbstractUploaderStrategy {
             }
         }
 
-        if (file instanceof String && !((String) file).matches("ftp:.*|https?:.*|s3:.*|data:[^;]*;base64,([a-zA-Z0-9/+\n=]+)")) {
+        if (file instanceof String && !((String) file).matches("(?s)ftp:.*|https?:.*|s3:.*|data:[^;]*;base64,([a-zA-Z0-9/+\n=]+)")) {
             file = new File((String) file);
         }
         String filename = (String) options.get("filename");
@@ -129,7 +129,7 @@ public class UploaderStrategy extends AbstractUploaderStrategy {
             response.close();
         }
 
-        if (code != 200 && code != 400 && code != 500) {
+        if (code != 200 && code != 400 && code != 404 && code != 500) {
             throw new RuntimeException("Server returned unexpected status code - " + code + " - " + responseData);
         }