+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+cloudinary-parent-1.4.4 / 2016-09-15
+====================================
+  * Fix issue when uploading URL with \n
 
 cloudinary-parent-1.4.3 / 2016-09-09
 ====================================